worker/requirements.txt
redis
rq
python-dotenv
boto3
asyncpg
sqlalchemy
